requirements.txt
Django<3.2
zappa<0.60
python-dotenv
psycopg2-binary
django-cors-headers
boto3<1.17
mail-parser<3.13
django_s3_storage
djangorestframework<3.12.4
django-extensions
